+++ ChangeLog.md
@@ -4,7 +4,9 @@
 
 * Change SIMD base64 codec to https://github.com/lemire/fastbase64.
 * Add SIMD enabled `count` from bytestring.
+* Switch `no-avx` on by default to improve compiled binary's portablity.
 * Add `collate`, courtesy by jgm: https://github.com/jgm/unicode-collation.
+* Add `hexDecodeWS/hexDecodeWS'` to `Z.Data.Vector.Hex`.
 
 ## 0.8.5.0  -- 2021-06-04